requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aws-requests-auth==0.4.3
boto3==1.38.18
botocore==1.38.18
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
deprecated==1.2.18
dill==0.4.0
docstring-parser==0.15
googleapis-common-protos==1.70.0
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.6.1
jmespath==1.0.1
log-symbols==0.0.14
markdown-it-py==3.0.0
mcp==1.9.0
mdurl==0.1.2
mpmath==1.3.0
ollama==0.4.8
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
pillow==11.2.1
pip==25.1.1
prompt-toolkit==3.0.51
protobuf==5.29.4
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
rich==14.0.0
s3transfer==0.12.0
six==1.17.0
slack-bolt==1.23.0
slack-sdk==3.35.0
sniffio==1.3.1
spinners==0.0.24
sse-starlette==2.3.5
starlette==0.46.2
strands-agents==0.1.1
strands-agents-builder==0.1.0
strands-agents-tools==0.1.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
